mumble.pro: don't add to QTPLUGIN if we're on a Qt with auto-plugin support.
Qt 5.3 introduced a feature where it'll automatically fill out QTPLUGIN
with appropriate auto-detected plugins for TEMPLATE=app.

This commit avoids adding entries to QTPLUGIN if we detect we're on a
Qt version that automatically fills out QTPLUGIN.

For Windows static builds, we don't use TEMPLATE=app, so for that
platform, we have to fill out QTPLUGIN in any case.