os_win: remove MumbleHWNDForQWidget, add mumble_mw_hwnd.
This commit removes MumbleHWNDForQWidget -- the only user of
private Qt API in Mumble on Windows.

It is replaced with a variable, mumble_hw_hwnd, which gets
filled out in main.cpp just after creating MainWindow.

Most MumbleHWNDForQWidget() uses were for g.mw, so the change
makes sense. The ones that weren't have been changed to use it,
or removed. Most of the removed instances are g.ocIntercept-related,
which is already broken beyond repair in Qt 5.