Audio, AudioInput, AudioOutput: add proper documentation about AudioInput and AudioOutput construction and destruction.
This commit documents the intent behind Audio::stopInput(), Audio::stopOutput() and Audio::stop().
Namely that the functions are coded the way they are to guarantee deterministic destruction of
AudioInput/AudioOutput on the main thread.

It also adds documentation to the AudioInput and AudioOutput constructors and destructors that
spells out the guarantee that they can only ever be called on the main thread, and only by
the Audio::startInput(), Audio::startOutput(), Audio::start(), Audio::stopInput(), Audio::stopOutput()
and Audio::stop() methods.