Revert PR #2674: Mumble: convert AudioInputPtr and AudioOutputPtr to use QSharedPointer instead of boost::shared_ptr.
This reverts PR #2674. That PR has a bug in stop(), stopInput() and stopOutput().

Specifically, the original code was structured such that destruction of
the AudioInput and AudioOutput objects would happen inside any of the stop methods.
However, the new implementations of stop(), stopInput() and stopOutput() in PR #2674
could cause the AudioInput and AudioOutput objects to be destroyed at any site that
takes a copy of g.ao or g.ai.

This happened in the PulseAudio backend. The AudioInput destructor would be called
while holding the PulseAudio mainloop lock. The destructor of the Pulse backend's
AudioInput also, tries to take the PulseAudio mainloop lock, causing an abort() to
happen.

Let us revert PR #2674 for now. There is no obvious way to implement the controlled
destruction behavior of the boost::shared_ptr with QSharedPointer. We also don't have
any code in Mumble that requires the audio shared pointers to be atomic (which was the
reason for the PR in the first place).

Fixes mumble-voip/mumble#2745