python.pri: detect missing 'which' command and try to recover.
Previously, the python.pri helper would try to query which python
binary on the system to use via the 'which' command.

Some systems do not have which (some Docker-based CI systems,
for example).

We now detect the presence of 'which'. If it is not available,
we try to use 'python' as the Python command. We also print a
warning about 'which' not being available -- as well as a message
pointing users to the MUMBLE_PYTHON environment variable.

Fixes mumble-voip/mumble#2748