compiler.pri: for win32-msvc, ignore C4091 when building against v140_xp toolset.
The Windows 7.1A SDK that is used when building with the v140_xp toolset
triggers the C4091 warning in multiple places. Instead of littering our
headers with pragmas, we ignore the warning everywhere via command-line
flag.

It shouldn't affect code quality. The x64 build doesn't ignore the warning
so we're bound to catch any warnings introduced in the future.