mumble_pch.hpp: Delete _WINSOCKAPI_ definition
The _WINSOCKAPI_ definition was needed when "winsock2.h" was being included after "windows.h", which checks if _WINSOCKAPI_ is defined to know if the header is included.
If it isn't, "winsock.h" (the previous version of winsock2.h) is included, which leads to a conflict and potential compilation fail, if the updated version of the header is included later.