AudioInput: limit echo canceller queue to 1 second of audio.
This sets an upper bound on the echo canceller's queue to avoid
situations like mumble-voip/mumble#1074 where a bug in PulseAudio
causes our echo queue to grow unbounded.

This fixes the memory leak aspect of mumble-voip/#1074, but echo
cancellation will be flakey in such setups.

We don't know if there is a workaround for the PulseAudio bug, so
let us do this for now.