LCD: drop workaround for Qt >= 5.6.2.
This drops the workaround we added in c04aa4e8 (see also
mumble-voip/mumble#2429). There was a Qt regression in
early 5.6 versions that caused our previously working
monochrome bitmap drawing code to break.

To be fair, our original code was not exactly correct
per the Qt docs, but it did produce correct results before
Qt 5.6.

Qt accepted the bug as a regression, and restored the old
behavior, so we can now keep using the old behavior.

We should re-visit this soon, because the drawing code
isn't actually correct according to the Qt docs.

But it works. At least until Qt 6.

Fixes mumble-voip/mumble#2815