Always use lrelease binary from QT_INSTALL_BINS.
Sometimes, Qt packages in Linux distros and the like provide an 'lrelease' binary in $PATH.
That is what we've relied on in mumble.pro previously.

However, that doesn't always work for all Qt installations.

To fix that, this change introduces a new lookup strategy for the 'lrelease' binary.
Instead of relying on it being in the user's default $PATH, we use qmake's QT_INSTALL_BINS
variable. That path is guaranteed to contain an 'lrelease' binary if the correct packages are
installed -- for both Qt 4 and Qt 5.