OverlayConfig: handle removal of qWinAppInst() in Qt 5.8.
This commit fixes the OverlayConfig code to build on Qt 5.8,
where qWinAppInst() was removed.

It was removed via
https://github.com/qt/qtbase/commit/64507c7165e42c2a5029353d8f97a0d841fa6b01

Thankfully, qWinAppInst() was very simple: it just called through
to GetModuleHandle(0) -- in both Qt 4, and Qt 5.

To sidestep the removal of qWinAppInst(), we fix our code by calling
GetModuleHandle() directly. (With 0 changed to NULL, which is what
we usually do in our coding style.)