Server, Cert: add Server::privateKeyFromPEM() method.
This adds a new method, Server::privateKeyFromPEM(), a helper
method for loading private keys.

In various places throughout the codebase, we use the same
sequence of operations for loading private keys from a PEM
bytestream.

This method moves that sequence to a single method, to avoid
the unnecessary duplication that is currently going on.