Register: treat certs from murmur.ini's sslCA option as CA certs when registering.
The sslCA option is used by people to specify extra CA certificates.

Before this patch series, the certificates from sslCA would be added
to the QSslSocket::defaultCaCertificates() list via MetaParams. Thus,
those certificates were automatically treated by QSslSocket as valid
root certificates -- by all QSslSocket instances in Murmur.

Since we've now stopped adding the sslCA certificates to the global
QSslSocket::defaultCaCertificates() list, we now have to add them to
each QSslSocket as necessary, for them to be treated as CA certificates.

One use-case for murmur.ini's sslCA option is to add extra CA
certificates in case the system CA store doesn't contain the necessary
certificates.

In Murmur, one of the only places CA certificates are used, is when
connecting to the registration server via HTTPS. In theory, the system
certificate store should contain the root CAs necessary for connecting
to the registration server on our supported OSes. But it might not for some
unsupported OSes. For example, if someone is using Murmur on an unsupported
OS where we or Qt do not know how to read the sytem's CA store, it will not
be able to establish a conection to the registration server. On such a system,
a user will be able to use the sslCA option in murmur.ini to fix certificate
validation for the registration server, by pointing sslCA to the system's
PEM certificate bundle, or a PEM copy of the Mozilla certificate store.

This commit keeps that behavior. Some people are possibly using sslCAs
for pointing Murmur to a CA bundle -- and we don't want to break their
setups.