mumble: use SHA-256 keys when storing self-generated blobs.
This updates the Mumble client to use SHA-256 instead of
SHA-1 when it eagerly caches blobs it generates itself.

When the Mumble client itself creates a blob, such as

 - A channel description
 - A user comment
 - A user textures
 - A channel comment

it will now SHA-256 hash the content and eagerly store it
in its blob database in a content-addressable manner.

Previously, it used a SHA-1 hash for the content-addressed
key.

For Murmur servers that use SHA-1 for blobs, this eager
caching will not work, since Murmur will still advertise
blobs via their SHA-1 hash. So, the next time the server sends
such a blob back to the client, the client will have to download
it again and put the SHA-1 variant into the blob database.