AudioOutputSpeech: only process Speex packets as Speex.
Otherwise, a malformed packet  could lead to a crash.

Note that this only affects packets generated by Mumble itself,
such as local loopback packets and voice recorder packets.

Packets from the network are filtered in such a way that a
packet with an invalid type will never be able to enter the
audio subsystem -- it is discarded in the network subsystem.