Audio, AudioOutput: guard against invalid packet types in AudioOutput::addFrameToBuffer().
For packets we receive from Murmur, we check that they're a valid voice
packet on receipt.

However, some packets are not received by Murmur, but generated by the
client itself: packets for RecordUser, and local loopback packets.

If we mess up their type due to a bug, we'd previously allow them into the
AudioOutput system. This commit discards them before they get there.