murmur: use SHA-256 keys for blobs in protocol messages.
This changes Murmur to return SHA-256 hashes for large blobs in protocol
messages. Previously, Murmur used SHA-1.

This has no real effect on clients. Clients trust the content-addressable
keys returned by Murmur for its content. No matter which hash function the
server is using, it will use the server-advertised keys in its own
blobstore. Thus, it is a fairly harmless operation to change the
hash-function used for blobs keys. The only thing it will trigger is that
existing blobs might need to be re-downloaded once.