Mumble: use WindowModal for all modal dialogs.
By default, Qt uses ApplicationModal for all QDialogs
when you exec() them.

However, in Mumble, some windows are in a separate window hierarchy,
and it would be nice to be able to switch to them even when a modal
dialog is shown on top of Mumble's MainWindow. These windows are the
Developer Console, and also the configuration dialog for ManualPlugin.

This commit tries to remedy that by marking all of Mumble's modal
windows as WindowModal instead of AppliationModal.

Note: as of now, this commit is not exhaustive, but it covers most
of the obvious cases. Right now, we can't set the window modality
for QMessageBoxes create via QMessageBox's static factory methods.
We'll probably have to provide our own MUMessageBox (or similar)
for that.