Cert: reset Server's SSL state in initializeCert().
This changes Server::initializeCert() to always reset the
Server's SSL state before reloading it.

Previously, we didn't reset the state.

Not doing so *does* actually works for the case where we
load the certificate and key from the per-vserver configuration
in ServerDB.

However, it doesn't work when reading from Meta, because
the old Server::initializeCert() would only use the settings
from Meta if qscCert/qskKey were null.