MurmurIce: base64-encode MurmurUser::context on the wire to avoid NUL bytes w/o losing data.
We now base64-encode the user's plugin context.
It is a binary blob, not generally useful for RPC users.

The most important property is the ability to detect that two
users are in the same game context -- and that property is preserved.

If the exact byte seqeuence is needed, it is possible to base64-decode
the returned string to retrieve it.

Fixes mumble-voip/mumble#1874