Settings: add setting for winhooks.
The setting is already used in GlobalShortcut_win.
We added it as a hidden setting because in some situations,
the winhooks engine is not preferable.

The code in GlobalShortcut_win already works -- so
this change is all we need to be able to easily access
the setting from the GlobalShortcut config dialog.