plugins/sto: Revamp plugin
This commit changes the code so that it follows the style of the other plugins.

The offsets are not updated, this means that the plugin only works with a very old version of the game, but with the new code it should be much more easy for a contributor to update them.

The commit also gets rid of a strict-aliasing rules break warning, which appeared when compiling with MinGW, caused by:

if (* reinterpret_cast<DWORD *>(& posblock[52]) == 1) {

because we're reading a spawn state or magical value, which is an unsigned char, from an array of floats.