scripts/rcc-depends.py: fall back to using the absolute path when os.path.relpath() fails on Windows.
On Windows, it is not always possible to create a relative path from an
absolute path. For example, if Qt lives on C:, and Mumble is being built
on Z:.

Things will fall apart then, because we include some of Qt's
translations in our .qrc files.

This commit works around that issue by falling back to absolute paths
when finding a relative path fails.