OverlayConfig: add UI for configuring the launcher filter.
This change adds the user interface and config dialog logic
needed for the overlay launcher filter.

This adds an additional mode to the overlay called the 'launcher filter'.

The idea behind the launcher filter is that we provide a sensible
whitelist of allowed launchers (which are typically digital storefronts,
and other launchers -- Steam, Origin, GOG, Battle.net, MMORPG launchers,
etc.), as well as a default whitelist of normal programs, because some
things don't fit into the 'launcher' check -- such as Guild Wars 2.

The existing 'blacklist' and 'whitelist' modes are still available,
as before.

This commit also changes the way "default" entries are shown.
Instead of being in a disabled state, they are now shown in italic.
Also, default entries now have a checkbox next to them, allowing users
to disable default entries.

Regarding new whitelist functionality:

To support the launcher filter overlay mode, we've added a default set of
programs to Mumble's whitelist.

However, when in 'whitelist' mode, Mumble has traditionally not provided
any defaults.

Also, for the whitelist mode to be useful to people, it should not have
any defaults -- because the whole purpose of the whitelist is for people
to be 100% in control themselves.

This commit also changes the UI in OverlayConfig to not show the defaults
from overlay_whitelist.h when in WhitelistExclusionMode.