TestTimer: make resolution test independent of QTime.
Previously, this test depended on the fact that QTime is/was a low
resolution timer, which is/may not always be the case.

This changes the test to only use the Timer class.

The test is re-written to run for 150 ms, to ensure that
we have enough samples to do a proper calculation. This
is also done to avoid fluctuations that could happen due
to context switches/overhead from other processes on the
test runner machine.

Since we don't compare to another timer anymore, we instead
explicitly test that the timer resolution is at least 5
microseconds. Currently, the Timer class doesn't spell out
its actual resolution, because its implementation varies
depending on platform.