MainWindow: avoid floor/ceil in VolumeUp/VolumeDown global shortcut handlers.
Previously, the code tried to use floorf (in VolumeUp) and ceilf (in
VolumeDown) in order to ensure that VolumeUp and VolumeDown rounded to
the nearest 10%.

However, with /fp:fast on MSVC, that didn't work in all cases. For
example, in VolumeDown, the volume was previously always calculated via
ceilf(). On some occasions, a previous invocation of VolumeDown will have
set g.s.fVolume to a value close to 1.3 (130%). This value, when passed to
ceilf() would become 1.4 (140%) and subtracting 10% from that would still
be 130%. Not as we would expect.

This commit fixes the logic to not expect on exactness of floats via
floorf/ceilf. Instead of rounding to the nearest 10%, we just subtract
or add 10% to the current volume. It doesn't matter that it isn't a
multiple of 10.