Backport PR #3033 to 1.2.x: compiler.pri: add CONFIG(dpkg-buildflags).
Shelling out to dpkg-buildflags to get the preferred build flags
is now the way to go on Debian.

Previously, our PPA packages used hardening-wrapper to pass the
proper hardening flags to the package. That package is no longer
available in Debian unstable or Ubuntu zesty. To remedy that situation,
we have to add flags ourselves via dpkg-buildflags.

This commit implements a new CONFIG flag, CONFIG(dpkg-buildflags)
which will query dpkg-buildflags and append the retrieved CFLAGS,
CXXFLAGS, CPPFLAGS and LDFLAGS to QMAKE_CFLAGS, QMAKE_CXXFLAGS,
QMAKE_CPPFLAGS, QMAKE_LDFLAGS to ensure they are used in the Mumble
build.