selfSignedServerCert_SHA1_RSA_2048: use RSA_generate_key_ex().
This replaces selfSignedServerCert_SHA1_RSA_2048()'s use of
the old, deprecated RSA_generate_key() function with
RSA_generate_key_ex().

The new function is a bit more explicit, so the code is a little
longer.

The RSA factor RSA_F4 is replaced with an explicit BIGNUM of 65537.
RSA_F4 is the flag to use for 65537 when using the old RSA_generate_key()
function.