TextToSpeech_unix: make setVolume not initialize speech-dispatcher.
It turns out that TextToSpeech attempts to set the volume on its engines
near initialization time.

This made our lazy-initialization pretty much useless.

This commit tries to remedy that by storing the last value passed to setVolume.
It then uses that stored value the first time speech-dispatcher is initialized.

As a result, we only initialize speech-dispatcher in say(), not setVolume().