src/tests: remove fragile TestTimer tests.
This commit drops the tests in TestTimer that depend on comparisons
to QTime to work.

In some environments, such as CI VMs/containers, we don't seem to get
the resolution we desire from QTime.

It's also worth noting that these tests were dormant in the repository
for a long time, until we hooked up 'make check'. So, they were quite
old, and assumptions made when the code was written might not hold up
any longer.

Fixes mumble-voip/mumble#3060