AudioInput, Settings: add 'audio/inputchannelmask' setting for selecting which mic channels should be mixed.
This adds a hidden option to allow users to select which microphone
channels will be downmixed to mono.

Examples of using the setting:

	audio/inputchannelmask = 0x0  # Use no channels. Nothing is recorded.
	audio/inputchannelmask = 0x1  # Only use left channel (channel index 0)
	audio/inputchannelmask = 0x2  # Only use right channel (channel	index 1)
	audio/inputchannelmask = 0x3  # Use both left and right channel.

The value of inputchannelmask is used as a bit mask. If a given bit is
set, the channel with the same index as the bit will be included in the
final downmix.

Fixes mumble-voip/mumble#3064