WASAPI, Settings: expose 'wasapi/role' setting to allow users to set WASAPI role.
By default, Mumble will open WASAPI devices with the eCommunications role,
since we're a voice chat applicaiton.

Some users have requested this to be configurable.

This commit adds a setting, 'wasapi/role', which is a string that maps
directly to the ERole enum from WASAPI:
https://msdn.microsoft.com/en-us/library/windows/desktop/dd370842

Users can now set 'wasapi/role' to one of 'console', 'multimedia' or
'communications'.

The default is still 'communications'.