TestServerResolver: modify test to work on systems that do not support both IPv4 and IPv6.
The initial test implementation was written targetting systems
that have both IPv4 and IPv6 available.

Qt's resolver does not return IPv6 results on systems that cannot
access the IPv6 internet. Nor does it return IPv4 results for systems
that cannot access the IPv4 internet.

Because of this, we modfiy the test case to be a little more
accommodating.