TestServerResolver: make the test Qt 4 compatible.
Qt 4 doesn't have QSignalSpy::wait().

This commit replaces calls to wait() with a similar construct
that works on Qt 4.

The implementation of QSignalSpy::wait() is quite simple:
It has its own QTestEventLoop and calls enterLoopMSecs()
on it. The return value of wait() is true if the count()
of the QSignalSpy changed since the last invocation.

Our construct is equivalent that that. Tests now pass on Qt 4.