ServerHandler: ensure only a single connection timeout timer is active at one time.
In mumble-voip/mumble#3153, it was reported that the new SRV code can
disconnect you from a server after a short while.

It turns out that we weren't being careful enough about our connection
timeout timer. A new timer was created for each connection attempt to
a server in the SRV list. In practice, this meant that 30 seconds after
each connection attempt, ServerHandler would try to tear down the
connection.

We now ensure we destroy the timer after we disconnect from a server.
This should ensure there is only one such timer active at any point in
time.

Fixes mumble-voip/mumble#3153