Settings: add DEPRECATED macro for keeping old, reserved names around.
This commit adds a new macro to Settings. It's called DEPRECATED, and it
is meant to be used to mark removed settings as deprecated.

It is primarily meant as documentation for developers, to ensure we don't
re-use a name in the future.

For loading, it's a no-op.

For saving, it removes the key from the user's settings store.